Implement lox.Parser
This is a recursive-descent parser which is able to parse and adhere to
the grammar rules of the lox programming language.

Currently it only supports expression statements and discards any tokens that are
not part of an expression statement.

It also handles syntax errors, although rudimentarily (missing
parenthesis in group statements, unterminated strings)